+++ cluster-network.tf
@@ -36,6 +36,7 @@ resource "oci_core_cluster_network" "cluster_network" {
   freeform_tags = {
       "cluster_name" = local.cluster_name
       "parent_cluster" = local.cluster_name
+      "user" = var.tags
   }
   placement_configuration {
     availability_domain = var.ad