+++ compute-cluster.tf
@@ -14,5 +14,6 @@ resource "oci_core_compute_cluster" "compute_cluster" {
   freeform_tags = {
     "cluster_name"  = local.cluster_name
     "parent_cluster" = local.cluster_name
+    "user" = var.tags
   }
 }