+++ controller.tf
@@ -103,6 +103,7 @@ resource "oci_core_instance" "controller" {
 resource "null_resource" "controller" { 
   depends_on = [oci_core_instance.controller, oci_core_volume_attachment.controller_volume_attachment ] 
   triggers = merge({
+    slurm_user_tags_enabled = tostring(var.slurm && var.slurm_user_tags_enabled)
     controller = oci_core_instance.controller.id
   }, var.slurm_job_notifications_enabled ? {
     slurm_notification_payload = local.slurm_notification_payload_hash
@@ -242,6 +243,7 @@ resource "null_resource" "cluster" {
     oci_identity_policy.slurm_notification_controllers,
   ]
   triggers = merge({
+    slurm_user_tags_enabled = tostring(var.slurm && var.slurm_user_tags_enabled)
     cluster_instances           = join(", ", local.cluster_instances_names)
     cluster_instance_ids        = join(", ", local.cluster_instances_ids)
     local_block_volume_settings = sha256(jsonencode([var.use_local_block_volume, var.local_block_volume_size, var.local_block_volume_performance, var.local_block_volume_mount_point]))
@@ -296,6 +298,7 @@ resource "null_resource" "cluster" {
       cluster_network = var.cluster_network,
       use_compute_agent = var.use_compute_agent,
       slurm = var.slurm,
+      slurm_user_tags_enabled = var.slurm && var.slurm_user_tags_enabled,
       slurm_job_notifications_enabled = var.slurm_job_notifications_enabled,
       slurm_notification_region = var.region,
       slurm_notification_compartment_id = var.targetCompartment,
@@ -457,6 +460,7 @@ resource "null_resource" "cluster" {
       scratch_nfs_path = var.scratch_nfs_path,
       use_scratch_nfs = var.use_scratch_nfs,
       slurm = var.slurm,
+      slurm_user_tags_enabled = var.slurm && var.slurm_user_tags_enabled,
       slurm_job_notifications_enabled = var.slurm_job_notifications_enabled,
       slurm_notification_region = var.region,
       slurm_notification_compartment_id = var.targetCompartment,